hot fix error handling for download DBGap sheet

+++ services/submission.js
@@ -1900,9 +1900,11 @@ class Submission {
             throw e;
         }
         finally {
-            const downloadDir = path.dirname(zipFile);
-            if (downloadDir && fs.existsSync(downloadDir)) {
-                fs.rmSync(downloadDir, {recursive: true, force: true });
+            if (zipFile && fs.existsSync(zipFile)) {
+                const downloadDir = path.dirname(zipFile);
+                if (downloadDir && fs.existsSync(downloadDir)) {
+                    fs.rmSync(downloadDir, {recursive: true, force: true });
+                }
             }
         }
     }
